Refactor: split openai conversion tests from stream variants
